Fix neovim opening in visual mode
Which key appears to partially execute the bound function and in this
case was running the `select_cell` command for jupynium keybindings.
This would result in the mode being set to VISUAL and not reset.
Thankfully these keybindings I had not changed so I added the default
bindings from Jupynium and have ignored the issue for now.